[cmake] Fix includes for Kodi headers
KODI_INCLUDE_DIR does already contain the 'kodi' directory. Therefore
specifying it in includes only works when the headers are in a compiler
default search path. See discussion in: xbmc/xbmc#9804

This allows to build the addon directly against a local version of
Kodi (depends on enet, moonlight-common-c and pugixml being installed
system wide). This is described in
https://github.com/xbmc/xbmc/blob/master/project/cmake/README.md#building-binary-addons.